Added anonymous user create page markup
