+++ prisma/schema.prisma
@@ -58,3 +58,6 @@ model VerificationToken {
 
   @@unique([identifier, token])
 }
+
+model Event {
+}
